pyproject.toml
[project]
name = "telegram-summarizer"
version = "0.1.0"
description = "Telegram bot that summarizes voice messages using local models"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "ollama>=0.1.0",
    "pydub>=0.25.1",
]

[project.scripts]
telegram-summarizer = "telegram_summarizer.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_summarizer"]
